CR addressed

+++ test/helper/Puppeteer_test.js
@@ -10,10 +10,12 @@ const AssertionFailedError = require('../../lib/assert/error');
 const formContents = require('../../lib/utils').test.submittedData(path.join(__dirname, '/../data/app/db'));
 const expectError = require('../../lib/utils').test.expectError;
 const webApiTests = require('./webapi');
+const FileSystem = require('../../lib/helper/FileSystem');
 
 let I;
 let browser;
 let page;
+let FS;
 const siteUrl = TestHelper.siteUrl();
 
 describe('Puppeteer', function () {
@@ -22,11 +24,15 @@ describe('Puppeteer', function () {
 
   before(() => {
     global.codecept_dir = path.join(__dirname, '/../data');
-    // create directory /download;
-    fs.mkdir(path.join(`${__dirname}/download`), () => {
+    // create download folder;
+    fs.mkdir(path.join(`${__dirname}/../data/download`), () => {
 
     });
-    global.output_dir = path.join(`${__dirname}/download`);
+    global.output_dir = path.join(`${__dirname}/../data/download`);
+
+    FS = new FileSystem();
+    FS._before();
+    FS.amInPath('download');
 
     I = new Puppeteer({
       url: siteUrl,
@@ -63,9 +69,11 @@ describe('Puppeteer', function () {
       if (err) throw err;
 
       for (const file of files) {
-        fs.unlink(path.join(global.output_dir, file), (err) => {
-          if (err) throw err;
-        });
+        if (file.includes('.mp4')) {
+          fs.unlink(path.join(global.output_dir, file), (err) => {
+            if (err) throw err;
+          });
+        }
       }
     });
   });
@@ -578,12 +586,19 @@ describe('Puppeteer', function () {
     });
   });
 
-  describe.only('#downloadFile', () => {
+  describe('#downloadFile', () => {
     it('should dowload file', async () => {
       await I.amOnPage('http://file-examples.com/index.php/sample-video-files/sample-mp4-files/');
-      const outputFile = await I.downloadFile('td[class="text-right file-link"] a');
+      const fileName = await I.downloadFile('td[class="text-right file-link"] a');
 
-      assert.equal(fs.existsSync(path.join(outputFile)), true);
+      FS.seeFile(fileName);
+    });
+
+    it('should dowload file with custom name', async () => {
+      await I.amOnPage('http://file-examples.com/index.php/sample-video-files/sample-mp4-files/');
+      const fileName = await I.downloadFile('td[class="text-right file-link"] a', 'thisisacustomname');
+
+      FS.seeFile(fileName);
     });
   });
 });
